hasher: avoid breaking system by qemu-user for current architecture
Installing qemu-user for the native architecture (say, qemu-user-aarch64
on aarch64) and registering it as a binfmt handler introduces an infinite
recursion (for any aarch64 binary gets executed via binfmt-misc now, and
binfmt-misc tries to execute qemu-user-aarch64, which is a aarch64 binary,
so it gets executed via binfmt-misc, which tries to execute
qemu-user-aarch64, and so on). Just DON'T do it.